backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
PyJWT>=2.8.0
smartapi-python>=1.4.8
pyotp>=2.9.0
python-dotenv>=1.0.0
httpx>=0.26.0
websockets>=12.0
logzero>=1.7.0
redis>=5.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
sqlalchemy[asyncio]>=2.0.0
uvloop>=0.19.0
gunicorn>=21.2.0
prometheus-fastapi-instrumentator>=6.1.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
xgboost>=2.0.0
apscheduler>=3.10.0
aiosqlite>=0.19.0
# ── ML Data Pipeline additions ──────────────────────
yfinance>=0.2.36
matplotlib>=3.8.0
joblib>=1.3.0
